Refactor feedback components: consolidate UserAvatar into FeedbackCard, add CompetencyChart, FeedbackHeader, FeedbackList, StatsCard, and StatsRow; remove unused UserAvatar file.
